Extend room admin api with additional attributes (#7225)
