+++ changelog.d/7225.misc
@@ -1,0 +1,1 @@
+Extend room admin api (`GET /_synapse/admin/v1/rooms`) with additional attributes.
